updated the function

+++ app/serializers.py
@@ -665,19 +665,19 @@ class TaskDetailsInputSerializer(serializers.Serializer):
 class AddProjectTimeSerializer(serializers.Serializer):
     project = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
     company_id = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
-    total_time = serializers.IntegerField(allow_null=False)
+    total_time = serializers.FloatField(allow_null=False)
     lead_name = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
     data_type = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
     editing_enabled = serializers.BooleanField(allow_null=False)
     
 class UpdateProjectTimeSerializer(serializers.Serializer):
     document_id = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
-    total_time = serializers.IntegerField(allow_null=False)
+    total_time = serializers.FloatField(allow_null=False)
 
 class UpdateProjectSpentTimeSerializer(serializers.Serializer):
     project = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
     company_id = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
-    spent_time = serializers.IntegerField(allow_null=False)
+    spent_time = serializers.FloatField(allow_null=False)
        
 class UpdateProjectTimeEnabledSerializer(serializers.Serializer):
     document_id = serializers.CharField(max_length=255,allow_null=False,allow_blank=False)
